moved react components to separate dir
